@@ -14,9 +14,9 @@
     "@angular/animations": "~7.2.0",
     "@angular/common": "~7.2.0",
     "@angular/compiler": "~7.2.0",
-    "@angular/core": "^7.0.2",
+    "@angular/core": "^7.2.14",
     "@angular/forms": "~7.2.0",
-    "@angular/http": "7.0.2",
+    "@angular/http": "7.2.14",
     "@angular/platform-browser": "~7.2.0",
     "@angular/platform-browser-dynamic": "~7.2.0",
     "@angular/router": "~7.2.0",
