@@ -1,7 +1,7 @@
-import { HttpClientModule } from "@angular/common/http";
-import { ApiService } from "src/app/services/api.service";
-import { SweetAlert2Module } from "@sweetalert2/ngx-sweetalert2";
-import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { ApiService } from 'src/app/services/api.service';
+import { HttpClientModule } from '@angular/common/http';
+import { SweetAlert2Module } from '@sweetalert2/ngx-sweetalert2';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 
@@ -9,7 +9,6 @@ import { LayoutRoutingModule } from "./layout-routing.module";
 import { LayoutComponent } from "./layout.component";
 import { HeaderComponent } from "./header/header.component";
 import { SidebarComponent } from "./sidebar/sidebar.component";
-import { RegisterUserComponent } from "../components/register-user/register-user.component";
 import { NgbModalModule } from "@ng-bootstrap/ng-bootstrap";
 import { RegisterRestaurantComponent } from '../components/register-restaurant/register-restaurant.component';
 import { EditRestaurantComponent } from '../components/edit-restaurant/edit-restaurant.component';
@@ -19,7 +18,6 @@ import { EditRestaurantComponent } from '../components/edit-restaurant/edit-rest
     LayoutComponent,
     HeaderComponent,
     SidebarComponent,
-    RegisterUserComponent,
     RegisterRestaurantComponent,
     EditRestaurantComponent
   ],
@@ -33,6 +31,6 @@ import { EditRestaurantComponent } from '../components/edit-restaurant/edit-rest
     HttpClientModule
   ],
   providers: [ApiService],
-  entryComponents: [RegisterUserComponent, RegisterRestaurantComponent, EditRestaurantComponent],
+  entryComponents: [RegisterRestaurantComponent, EditRestaurantComponent],
 })
-export class LayoutModule {}
+export class LayoutModule { }
