@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { NgbModal, ModalDismissReasons } from '@ng-bootstrap/ng-bootstrap';
+import { GrupoOncePagoService } from 'src/app/services/grupo-once-pago/grupo-once-pago.service';
+
 
 @Component({
   selector: 'app-grupo-once-pago',
@@ -7,9 +10,38 @@ import { Component, OnInit } from '@angular/core';
 })
 export class GrupoOncePagoComponent implements OnInit {
 
-  constructor() { }
+  closeResult: string;
+  constructor(private modalService: NgbModal,
+ private _grupooncepagoSerivce: GrupoOncePagoService
+    
+    
+    ) { }
+
+  open(content) {
+      this.modalService.open(content).result.then((result) => {
+          this.closeResult = `Closed with: ${result}`;
+      }, (reason) => {
+          this.closeResult = `Dismissed ${this.getDismissReason(reason)}`;
+      });
+    }
+
+    private getDismissReason(reason: any): string {
+        if (reason === ModalDismissReasons.ESC) {
+            return 'by pressing ESC';
+        } else if (reason === ModalDismissReasons.BACKDROP_CLICK) {
+            return 'by clicking on a backdrop';
+        } else {
+            return  `with: ${reason}`;
+        }
+    }
+
+    public payMethods =[];
+    public orderList =[];
+
 
   ngOnInit() {
+      this.payMethods = this._grupooncepagoSerivce.getPaymentMethod();
+      this.orderList = this._grupooncepagoSerivce.getOrderList();
   }
 
 }
