@@ -60,13 +60,22 @@ const routes: Routes = [
       {
         path: 'grupo-once-pago',
         loadChildren:
-        '../components/grupo-once-pago/grupo-once-pago.module#GrupoOncePagoModule'
+          '../components/grupo-once-pago/grupo-once-pago.module#GrupoOncePagoModule'
       },
       {
         path: 'restaurantes',
         loadChildren:
-        '../components/restaurantes/restaurantes.module#RestaurantesModule'
+          '../components/restaurantes/restaurantes.module#RestaurantesModule'
+      },
+      {
+        path: 'users',
+        loadChildren: '../components/users/users-list/users-list.module#UsersListModule'
+      },
+      {
+        path: 'grupo-cinco',
+        loadChildren: '../components/grupo-cinco/grupo-cinco.module#GrupoCincoModule'
       }
+
     ]
   }
 ];
@@ -74,4 +83,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class LayoutRoutingModule {}
+export class LayoutRoutingModule { }
